DOC: remove pyplot from mpl API example
Create Axes objects and pass them into pandas to avoid registering
them with pyplot.  This avoids:

 - needing to close the figures to avoid memory leaks (as no one else
   is holding a reference to it)
 - needing to close the figure to avoid the `inline` backend
   automatically displaying the figure.

The second problem could be avoided by switch to a non-interactive
backend (such as Agg) but would not solve the problem of the memory
leak or getting the figure object implicitly created by pandas.