moving fixtures straight into conftest (#700)
* moving fixtures straight into conftest

enables autocompletion and gets rid of a
superfluous folder. also added two
fixtures to make the server tests a little
dry-er.

* Added test cases for state repr and kill_all_servers command

also fixed a bug in the server kill command. Bokeh raises an
AssertionError when calling `stop` on a server that was already stopped
through other means. This can safely be ignored in the kill_all routine,
but it wasn't as of yet.

* accessed and executed stops incorrectly

* remove panel type from repr test

it's in str in py3 and unicode in py2. Just removing it from the
assertion since it's not the important part anyways.

* rebase on master, apply fixes back