Add id to facilitate the styling of each card
besides changing all the card properties with
.react-grid-item{
   background-color: #545454 !important;

 }

you can do something like

#reactcard4 {border: 5px solid #6b5b95;
padding: 10px;
           padding-top: 20px !important;}

#reactcard3 {overflow: auto;}

This solve several problems besides the styling.

With the padding and depending in the card content, you can avoid that the maximize item gets overlaped with the bokeh logo. Besides with overflow in a specific card you can put items which are not completely responsive like tables.