Add ComponentResourceHandler to server (#3284)
This PR implements a so called ComponentResourceHandler that is always added to the server. The ComponentResourceHandler resolves paths relative to some module, i.e. it resolves patterns like <endpoint>/<module>/<class>/<attribute>/<path_to_resource>. Internally the resource handler then resolves the path to the module that was referenced, the class that was referenced and finally the attribute on that class. If and only if the requested resource is listed in that attribute will it resolve the path_to_resource and serve it. This means that ReactiveHTML and Template resources that reference local files can now be resolved correctly without having to be bundled into the Panel bundle itself.