Fix flawed authorize logic
The previous commit did not allow an authorized user to see the
requested page, if a `config.authorize_callback` was given due to faulty
logic. This commit moves page creation above all authorization checks,
and only modifies the page object if a `config.authorize_callback` is
given, and if the user (or user and requested path) are authorized.