Fix pandas warning when sorting filtered table
When a Tabulator with applied filter(s) is sorted the code yields a warning from pandas "A value is trying to be set on a copy of a slice from a DataFrame".
This changes fixes the ussie by first sorting Tabulator data and then applying filter. Alternative is to create a copy of passed dataframe to _sort_df function.