Update CrossSelector.ipynb
Hi. Not a very experienced Github user, Ihope I proposed the change in the right way (and that there is indeed an issue in the current documentation).

Line 31: says value of selected options is a boolean. That needs to be list instead though right? Value is stated as type list in similar selector widgets. Changed boolean to list in line 31 in the fork I made.

Also, the below seems unclear (for me). Defaults to re.search? Maybe insert a hyperlink to an explanation of re.search?

It states that Function has two arguments. Maybe insert or link to anexplanation / example(s) of a query with 2 arguments?

I get that you can put a (regex?) query in the search boxes, but doesn't that query search through  the list of item labels as shown  in the box?

What is the purpose of the 2nd argument 'item label'? Or does item label mean the labels in the search boxes that say 'Filter available options' and Filter selected options? A link or hover to what an item label is in this context would be nice.

 "* **``filter_fn``** (function): The filter function applied when querying using the text fields, defaults to re.search. Function is two arguments, the query or pattern and the item label.\n",
   
    "* **``value``** (list): Currently selected options\n",