Update other_nb.md
Added warning that ipywidgets + Panel does not work in Colab.

I think this should be pointed out in the documentation (I have wasted considerable time trying to get it to work, which is currently not possible due to Colab constraints regarding comms). 

Maybe even add a warning in the code that detects use of ipywidgets + Colab and provides a warning in the Cell output?

See Discord message by @jbednar and surrounding messages here: https://discord.com/channels/1075331058024861767/1088157184489164831/1206732316513271818

And Github issue here: https://github.com/holoviz/panel/issues/6298

Added mention of Anaconda Cloud as a Notebook environment, and mention that users should upgrade Panel when using it. I am not sure what the correct / optimal !pip install command is. The specific command should probably be added, but at least this gives users a heads-up.

Should Panel have code that checks its own version, and produces  a warning if  the latest version is not being used?

That would seem sensible, also to alert users in local environments that are currently on the latest version, but might not be when the next version comes out. So this would provide them with a heads-up.