Update FileInput.ipynb to add reference to FileDropper for uploading large files.
The current doc mentions FileInput is not meant for uploading large files. But it makes no mention of FileDropper yet, which does support this.  


The ``FileInput`` widget allows uploading one or more file from the frontend and makes the filename, file data and [MIME type](https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types) available in Python.

Added this:

It is not optimzed for uploading large files. Use the ``FileDropper`` widget instead if this is required.\n",
    "\n",