Tabulator custom filtering function returning mask instead of filtered dataframe (#7225)
This changes allows to pass a custom function for filtering tabulator data. The function should instead of returning new filtered dataframe return a mask that is then treated as other filter masks. This  change allows to combine multiple filters to a tabulator data when one or multiple filters are defined as custom functions.