Incorrect hover text for max_width and max_height: both said Minimum instead of Maximum
In the Layout Tab of .controls, when hovering over the ? info icon of Max width and Max height, the hover text says: Minimal width / height ... etc. This should be Maxium width / height ... etc.

So currently, the hover text for these 2 elements is identical to that of Min Width and Min height.

Changed "Minimal" to Maximum in the doc string for max_width and max_height.

+1 for .controls by the way! Great way to understand / iterate what various aspects of a component do.

And .controls itself is a nice example of what Panel can do.

Additional observation: the 'Visible' element at the bottom of the Layout tab does not have a ? icon + hover text yet.

Same for the 2 control elements ('value' and 'Disabled') in the first tab of .controls.