Update Checkbox.ipynb to indicate both Toggle and Switch are interchangeable with Checbox. (#7738)
Textual improvements to clarify all 3 widgets are interchangeable, and adds links to the 2 docs for the other widgets that are interchangeable. See #7737

No entirely sure if the added links are structured correctly, so that they will be converted to correct links when the webpage is generated from the Notebook.

See issue https://github.com/holoviz/panel/issues/7737 for request for further explanation on this.