Update Plotly.ipynb to include dynamic reference to current Plotly version
Now that https://github.com/holoviz/panel/pull/7447 has been merged, versions of libraries used by Panel can be referred to dynamically in the docs.

This PR is a first iteration of this. It adds the sentence:

Panel uses Plotly version x.

To the docs, by adding the code line:

 "Panel uses [Plotly](https://plotly.com/python/) **version {{PLOTLY_VERSION}}** \n",
    "\n",


Suggestions for alternate / improved phrasing are very welcome. Please chime in @philippjfr , @MarcSkovMadsen , @hoxbro if you have suggestions.

Once the optimal phrasing has emerged, the intention is to apply it to other reference docs, to ensure uniformity between docs.

The intention is  to submit those subsequent doc changes as 1 PR.